requirements.txt
aiohttp==3.13.3
asyncio==4.0.0
beautifulsoup4==4.14.3
davey==0.1.0
discord.py[voice]==2.7.0
ollama==0.6.1
PyNaCl==1.5.0
regex==2026.2.28
yt-dlp==2026.3.3
yt-dlp-ejs==0.5.0